refactor(capture): split parallel driver modules
